+++ src/main/java/net/silentchaos512/gear/traits/TraitConst.java
@@ -2,11 +2,13 @@ package net.silentchaos512.gear.traits;
 
 import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.gear.SilentGear;
+import net.silentchaos512.gear.api.traits.ITrait;
 
 public class TraitConst {
     // Traits needed somewhere in the code
     public static final ResourceLocation ANCIENT = SilentGear.getId("ancient");
     public static final ResourceLocation CRUDE = SilentGear.getId("crude");
+    public static final ResourceLocation FLAMMABLE = SilentGear.getId("flammable");
     public static final ResourceLocation INDESTRUCTIBLE = SilentGear.getId("indestructible");
     public static final ResourceLocation LUSTROUS = SilentGear.getId("lustrous");
     public static final ResourceLocation MAGMATIC = SilentGear.getId("magmatic");