+++ src/main/java/net/silentchaos512/gear/util/TextUtil.java
@@ -1,0 +1,16 @@
+package net.silentchaos512.gear.util;
+
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TranslationTextComponent;
+
+public final class TextUtil {
+    private TextUtil() {throw new IllegalAccessError("Utility class");}
+
+    public static ITextComponent translate(String prefix, String suffix) {
+        return new TranslationTextComponent(prefix + ".silentgear." + suffix);
+    }
+
+    public static ITextComponent translate(String prefix, String suffix, Object... args) {
+        return new TranslationTextComponent(prefix + ".silentgear." + suffix, args);
+    }
+}