+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Flammable trait (only on wood by default). Flammable items take damage if the player is on fire and the item is equipped. They can also be used as fuel in furnaces, with the burn time being proportional to the item's durability.
 ### Fixed
 - Parts GUI not displaying stats correctly [#128]